+++ microhapulator/tests/test_type.py
@@ -20,9 +20,8 @@ def test_type_simple():
     fasta = data_file('pashtun-sim/tiny-panel.fasta.gz')
     gt = microhapulator.type.type(bam, fasta)
     testgtfile = data_file('pashtun-sim/test-output.json')
-    testgt = microhapulator.genotype.ObservedGenotype(filename=testgtfile)
-    assert gt.data == testgt.data
-    assert gt.dump() == testgt.dump()
+    testgt = microhapulator.genotype.ObservedGenotype(fromfile=testgtfile)
+    assert gt == testgt
 
 
 def test_type_missing_bam_index():
@@ -36,13 +35,13 @@ def test_type_missing_bam_index():
 def test_type_cli_simple():
     with NamedTemporaryFile() as outfile:
         arglist = [
-            'type', '--out', outfile.name, data_file('pashtun-sim/tiny-panel.fasta.gz'),
+            'type', '--out', outfile.name, '--threshold', '5',
+            data_file('pashtun-sim/tiny-panel.fasta.gz'),
             data_file('pashtun-sim/aligned-reads.bam'),
         ]
         args = microhapulator.cli.parse_args(arglist)
         microhapulator.type.main(args)
         testgtfile = data_file('pashtun-sim/test-output.json')
-        with open(outfile.name, 'r') as fh, open(testgtfile, 'r') as testfh:
-            gtdata = json.load(fh)
-            testgtdata = json.load(testfh)
-            assert gtdata == testgtdata
+        gtdata = microhapulator.genotype.ObservedGenotype(fromfile=outfile.name)
+        testgtdata = microhapulator.genotype.ObservedGenotype(fromfile=testgtfile)
+        assert gtdata == testgtdata